NEW Custom map projections for DataPointsLayer
