+++ Widgets/Parts/Maps/DataShapesLayer.php
@@ -14,12 +14,11 @@ use exface\Core\Interfaces\Widgets\WidgetLinkInterface;
 use exface\Core\Widgets\DataColumn;
 use exface\Core\DataTypes\NumberDataType;
 use exface\Core\Widgets\Parts\Maps\Interfaces\GeoJsonMapLayerInterface;
-use exface\Core\Widgets\Parts\Maps\Interfaces\MapProjectionInterface;
-use exface\Core\Widgets\Parts\Maps\Projection\Proj4Projection;
 use exface\Core\Widgets\Parts\Maps\Interfaces\CustomProjectionMapLayerInterface;
 use exface\Core\Widgets\Parts\DragAndDrop\DropToAction;
 use exface\Core\Interfaces\Widgets\iCanBeDragAndDropTarget;
 use exface\Core\Widgets\Parts\Maps\Interfaces\GeoJsonWidgetLinkMapLayerInterface;
+use exface\Core\Widgets\Parts\Maps\Traits\CustomProjectionLayerTrait;
 
 /**
  * 
@@ -52,6 +51,8 @@ class DataShapesLayer extends AbstractDataLayer
     
     use ValueLabeledLayerTrait;
     
+    use CustomProjectionLayerTrait;
+    
     private $shapeAttributeAlias = null;
     
     private $shapeColumn = null;
@@ -67,10 +68,6 @@ class DataShapesLayer extends AbstractDataLayer
     private $lineWeight = null;
     
     private $opacity = null;
-    
-    private $projection = null;
-    
-    private $projectionConfig = null;
     
     private $valuePosition = self::VALUE_POSITION_TOOLTIP;
     
@@ -298,55 +295,6 @@ class DataShapesLayer extends AbstractDataLayer
         $widget = $this->initDataWidgetValue($widget);
         
         return $widget;
-    }
-    
-    /**
-     * 
-     * {@inheritDoc}
-     * @see \exface\Core\Widgets\Parts\Maps\Interfaces\CustomProjectionMapLayerInterface::getProjection()
-     */
-    public function getProjection() : MapProjectionInterface
-    {
-        return $this->projection;
-    }
-    
-    /**
-     * The projection to be used - e.g. `EPSG:25832` or `EPSG:3857` - or empty for autodetection.
-     * 
-     * Example:
-     * 
-     * ```
-     * {
-     *  "projection": {
-     *      "name": "EPSG:25832",
-     *      "definition": "+proj=utm +zone=32 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
-     *  }
-     * }
-     *  
-     * ```
-     *
-     * @uxon-property projection
-     * @uxon-type \exface\Core\Widgets\Parts\Maps\Projection\Proj4Projection
-     * @uxon-template {"name": "", "definition": ""}
-     *
-     * @param string $value
-     * @return DataShapesLayer
-     */
-    protected function setProjection(UxonObject $value) : DataShapesLayer
-    {
-        $this->projection = new Proj4Projection($value->getProperty('name'));
-        $this->projection->importUxonObject($value);
-        return $this;
-    }
-    
-    /**
-     * 
-     * {@inheritDoc}
-     * @see \exface\Core\Widgets\Parts\Maps\Interfaces\CustomProjectionMapLayerInterface::hasProjectionDefinition()
-     */
-    public function hasProjectionDefinition() : bool
-    {
-        return $this->projection !== null;
     }
     
     /**
